+++ packages/@vangware/iterables/src/take.ts
@@ -21,10 +21,19 @@ export const take = (amount: bigint | number) =>
 				// eslint-disable-next-line functional/no-let
 				let count = 0n;
 
-				// eslint-disable-next-line functional/no-loop-statements
-				for (const item of iterable) {
-					// eslint-disable-next-line @typescript-eslint/no-unused-expressions
-					count < amount ? (yield item, (count += 1n)) : undefined;
+				// eslint-disable-next-line functional/no-conditional-statements
+				if (amount > 0) {
+					// eslint-disable-next-line functional/no-loop-statements
+					for (const item of iterable) {
+						// eslint-disable-next-line @typescript-eslint/no-unused-expressions, functional/no-conditional-statements
+						if (count < amount) {
+							yield item;
+							// eslint-disable-next-line functional/no-expression-statements
+							count += 1n;
+						} else {
+							break;
+						}
+					}
 				}
 			},
 	)(
@@ -33,10 +42,19 @@ export const take = (amount: bigint | number) =>
 				// eslint-disable-next-line functional/no-let
 				let count = 0n;
 
-				// eslint-disable-next-line functional/no-loop-statements
-				for await (const item of iterable) {
-					// eslint-disable-next-line @typescript-eslint/no-unused-expressions
-					count < amount ? (yield item, (count += 1n)) : undefined;
+				// eslint-disable-next-line functional/no-conditional-statements
+				if (amount > 0) {
+					// eslint-disable-next-line functional/no-loop-statements
+					for await (const item of iterable) {
+						// eslint-disable-next-line @typescript-eslint/no-unused-expressions, functional/no-conditional-statements
+						if (count < amount) {
+							yield item;
+							// eslint-disable-next-line functional/no-expression-statements
+							count += 1n;
+						} else {
+							break;
+						}
+					}
 				}
 			},
 	) as <Iterable extends IsomorphicIterable>(